Use int for error-status and error-index
Change error-status and error-index from long to int in ReportPdu and Parser. Parser now casts ReadInteger results to int; ReportPdu constructor signature updated to accept ints. This aligns the types for SNMP error fields (small integers) and avoids unnecessary long usage.

Files changed:
- SapphTools.Snmp/Messages/Parser.cs
- SapphTools.Snmp/Pdu/ReportPdu.cs

No behavioral changes beyond the type narrowing; update callers if they relied on long values.

+++ SapphTools.Snmp/Messages/Parser.cs
@@ -141,8 +141,8 @@ public static class Parser {
                                     ReadOnlySpan<byte> pduBody) {
         int pos = 0;
         long requestId = ReadInteger(pduBody, ref pos, "request-id");
-        long errorStat = ReadInteger(pduBody, ref pos, "error-status");
-        long errorIndex = ReadInteger(pduBody, ref pos, "error-index");
+        int errorStat = (int)ReadInteger(pduBody, ref pos, "error-status");
+        int errorIndex = (int)ReadInteger(pduBody, ref pos, "error-index");
 
         ReadOnlySpan<byte> vbListRest = pduBody[pos..];
         Asn1Tag vbListTag = ReadTag(vbListRest);
